Add adjacency list
