+++ backend/app/scripts/__init__.py
@@ -1,0 +1,1 @@
+# Create module to import scripts
